backend/requirements.txt
absl-py==2.1.0
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.4.0
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.0
cssselect2==0.8.0
daphne==4.1.2
defusedxml==0.8.0rc2
dj-database-url==2.3.0
dj-stripe==2.9.1
Django==5.1.3
django-bootstrap==0.2.4
django-bootstrap5==24.3
django-cors-headers==4.6.0
django-filter==24.3
django-paypal==2.1
django-phonenumber-field==8.0.0
django-rest-framework==0.1.0
django-restframework==0.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
fonttools==4.58.0
git-filter-repo==2.47.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
immutabledict==4.2.1
incremental==24.7.2
iniconfig==2.1.0
msgpack==1.1.0
numpy==2.2.3
oauthlib==3.2.2
ortools==9.12.4544
packaging==24.2
pandas==2.2.3
paypalrestsdk==1.13.3
phonenumbers==8.13.52
pillow==11.1.0
pluggy==1.5.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydyf==0.11.0
PyJWT==2.10.1
pyOpenSSL==25.0.0
pyphen==0.17.2
pytest==8.3.5
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2025.1
redis==5.2.1
reportlab==4.4.1
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.2.0
setuptools==75.8.0
six==1.17.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
stripe==11.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
weasyprint==65.1
webencodings==0.5.1
zope.interface==7.2
zopfli==0.2.3.post1
